fixed output to have community_stats
