+++ api/src/main/java/com/codeforcommunity/dto/report/GetCommunityStatsResponse.java
@@ -1,27 +1,15 @@
 package com.codeforcommunity.dto.report;
 
+import java.util.List;
+
 public class GetCommunityStatsResponse {
-    private final Integer adopterCount;
-    private final Integer treesAdopted;
-    private final Integer stewardshipActivities;
+  public final List<CommunityStats> communityStatsReport;
 
-    public GetCommunityStatsResponse(Integer adopterCount,
-                             Integer treesAdopted,
-                             Integer stewardshipActivities) {
-        this.adopterCount = adopterCount;
-        this.treesAdopted = treesAdopted;
-        this.stewardshipActivities = stewardshipActivities;
-    }
+  public GetCommunityStatsResponse(List<CommunityStats> communityStatsReport) {
+    this.communityStatsReport = communityStatsReport;
+  }
 
-    public Integer getAdopterCount() {
-        return this.adopterCount;
-    }
-
-    public Integer getTreesAdopted() {
-        return this.treesAdopted;
-    }
-
-    public Integer getStewardshipActivities() {
-        return this.stewardshipActivities;
-    }
+  public List<CommunityStats> getcommunityStatsReport() {
+    return communityStatsReport;
+  }
 }
